requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
azure-ai-projects>=1.0.0b11
azure-identity>=1.14.0
azure-monitor-opentelemetry>=1.0.0
opentelemetry-sdk>=1.18.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
semantic-kernel
